test(util): move TestConvertNumbers into util_test.go
